@@ -85,6 +85,5 @@
 </table>
 <page-link page-model="Model.PageViewModel" page-action="Index"></page-link>
 <p>
-    <a asp-action=".../Indicators/Index" class="btn btn-sm btn-primary">Вернуться к показателям</a>
-
+    <a asp-controller="Indicators" asp-action="Index" class="btn btn-sm btn-primary">Вернуться к показателям</a>
 </p>
