@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using Microsoft.AspNetCore.Http.Authentication;
+//using Microsoft.AspNetCore.Http.Authentication;
 using Microsoft.AspNetCore.Identity;
 
 namespace GomelRectorCouncil.Models.ManageViewModels
@@ -8,6 +8,6 @@ namespace GomelRectorCouncil.Models.ManageViewModels
     {
         public IList<UserLoginInfo> CurrentLogins { get; set; }
 
-        public IList<AuthenticationDescription> OtherLogins { get; set; }
+        //public IList<AuthenticationDescription> OtherLogins { get; set; }
     }
 }
